# docs/api/v2/openapi.yaml
openapi: "3.2.0"
info:
  title: PicoCompute API
  version: "2.0.0"
  $self: https://api.picocompute.dev/v2/openapi.yaml
  description: |
    PicoCompute public API for sandbox lifecycle, command execution, file access,
    workspace streaming, and port exposure.

    ## Versioning

    This is the v2 API. Breaking changes from v1:

    - **State model**: 12 states (v1 had 15). `Executing` and `Idle` are
      removed. `Requested` renamed to `Pending`.
    - **Idempotency**: All lifecycle operations require an `Idempotency-Key`
      header.
    - **Tenant context**: All requests require a `PicoCompute-Tenant` header.
    - **Access leases**: Exec, stream, and file operations require a signed
      `PicoCompute-Access-Lease` header.
    - **Long-running operations**: Create, Suspend, Resume, and Destroy return
      an operation token for async polling.
    - **Response envelopes**: All responses include a `request_id`.

    ## Error taxonomy

    | Status | Code | Meaning |
    |--------|------|---------|
    | 400 | `validation_error` | Request schema or constraint violation |
    | 401 | `unauthenticated` | Missing or invalid bearer token |
    | 403 | `access_denied` | Lease expired, revoked, or scope mismatch |
    | 403 | `quota_exceeded` | Tenant resource quota exhausted |
    | 403 | `policy_denied` | Policy engine rejected the request |
    | 404 | `not_found` | Sandbox or resource does not exist |
...
    | 500 | `runtime_error` | Internal runtime failure |
...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /sandboxes:
    post:
      tags: [create, sandboxes]
      operationId: CreateSandbox
      summary: Create a new sandbox or fork an existing one
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Tenant"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSandboxRequest"
            examples:
              create:
                summary: Create a standard Firecracker sandbox
                value:
                  runtime: firecracker
                  image: alpine-linux-6.1
                  vcpus: 2
                  memory_mb: 512
                  idle_timeout_secs: 300
                  ports: [8080]
              fork:
                summary: Fork from an existing sandbox
                value:
                  source_sandbox_id: sbx_01JXYZABCDEFGH
                  idle_timeout_secs: 600
      responses:
        "202":
          description: Sandbox creation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
              examples:
                accepted:
                  value:
                    operation_id: op_01JXYZ123456789
                    sandbox_id: sbx_01JXYZABCDEFGH
                    state: Pending
                    status_url: /v2/operations/op_01JXYZ123456789
        "201":
          description: |
            Sandbox created synchronously (available for stub and
            pre-warmed runtimes). The idempotency-key replay result
            also uses this status code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/QuotaExceeded"
        "409":
          description: Idempotency key conflict (different body with same key)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: idempotency_conflict
                  message: A different request was already submitted with this idempotency key
                  request_id: req_01JXYZ123456789

    get:
      tags: [sandboxes]
      operationId: ListSandboxes
      summary: List sandboxes for the tenant
      parameters:
        - $ref: "#/components/parameters/Tenant"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 250
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
        - name: state
          description: Filter by lifecycle state
          in: query
          schema:
            $ref: "#/components/schemas/SandboxState"
        - name: runtime
          description: Filter by runtime type
          in: query
          schema:
            $ref: "#/components/schemas/RuntimeType"
      responses:
        "200":
          description: Paginated list of sandboxes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxList"
  /sandboxes/{sandbox_id}:
    get:
      tags: [sandboxes]
      operationId: GetSandbox
      summary: Get sandbox status and metadata
      parameters:
        - $ref: "#/components/parameters/SandboxId"
        - $ref: "#/components/parameters/Tenant"
      responses:
        "200":
          description: Sandbox found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [lifecycle, sandboxes]
      operationId: DestroySandbox
      summary: Destroy a sandbox and release its resources
      parameters:
        - $ref: "#/components/parameters/SandboxId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Tenant"
      responses:
        "202":
          description: Destroy accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "200":
          description: |
            Sandbox already destroyed (idempotent replay) or destroyed
            synchronously.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/StateConflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token for API authentication. Scoped to a tenant. Obtain from
        the PicoCompute control plane.
  parameters:
    SandboxId:
      name: sandbox_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^sbx_[a-zA-Z0-9]+$"
      example: sbx_01JXYZABCDEFGH
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 256
        pattern: "^[a-zA-Z0-9_-]+$"
      description: |
        Client-generated key for idempotent request replay. Scoped to the
        tenant. Results are retained for 24 hours.
    Tenant:
      name: PicoCompute-Tenant
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[a-z][a-z0-9_-]*$"
      description: Tenant identifier for multi-tenancy.
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: validation_error
              message: "sandbox_id must match pattern ^sbx_[a-zA-Z0-9]+$"
              details:
                - field: sandbox_id
                  issue: pattern_mismatch
                  expected: "^sbx_[a-zA-Z0-9]+$"
                  received: "my-sandbox"
              request_id: req_01JXYZ123456789
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: "sandbox sbx_01JXYZABCDEFGH not found"
              request_id: req_01JXYZ123456789
    QuotaExceeded:
      description: Tenant resource quota exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: quota_exceeded
              message: "Tenant sandbox limit (10) reached"
              details:
                limit: 10
                current: 10
                resource: sandbox_count
              request_id: req_01JXYZ123456789
...
    StateConflict:
      description: Sandbox state does not allow this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: state_conflict
              message: "Sandbox must be Running to exec (current state: Suspended)"
              details:
                current_state: Suspended
                required_states:
                  - Running
              request_id: req_01JXYZ123456789
...
  schemas:
...
    SandboxState:
      type: string
      enum:
        - Pending
        - Scheduled
        - Preparing
        - Booting
        - Running
        - Suspending
        - Suspended
        - Resuming
        - Stopped
        - Destroying
        - Destroyed
        - Failed
    RuntimeType:
      type: string
      enum:
        - firecracker
        - qemu
    FailureCode:
      type: string
      enum:
        - boot_timeout
        - resource_exhausted
        - network_unreachable
        - suspend_failed
        - resume_failed
        - workspace_corrupted
        - image_pull_failed
        - quota_exceeded
        - policy_denied
        - internal_error
    OperationStatus:
      type: string
      enum:
        - pending
        - in_progress
        - completed
        - failed
...
    CreateSandboxRequest:
      type: object
      required:
        - image
      properties:
        runtime:
          allOf:
            - $ref: "#/components/schemas/RuntimeType"
          default: firecracker
        image:
          type: string
          minLength: 1
          maxLength: 256
          description: Root filesystem image identifier
          example: alpine-linux-6.1
        source_sandbox_id:
          type: string
          pattern: "^sbx_[a-zA-Z0-9]+$"
          description: |
            When set, forks the workspace and configuration from this sandbox.
            The source sandbox state is unchanged by the fork.
          example: sbx_01JXYZABCDEFGH
        vcpus:
          type: integer
          minimum: 1
          maximum: 64
          default: 2
        memory_mb:
          type: integer
          minimum: 64
          maximum: 65536
          default: 512
        idle_timeout_secs:
          type: integer
          minimum: 0
          maximum: 86400
          default: 300
          description: Seconds of inactivity before automatic stop. 0 disables.
        ports:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 65535
          uniqueItems: true
          description: Guest ports to expose on the host
          example: [8080, 3000]
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables to inject into the sandbox
        labels:
          type: object
          additionalProperties:
            type: string
          description: Arbitrary key-value metadata for the sandbox
...
    Sandbox:
      type: object
      properties:
        id:
          type: string
          example: sbx_01JXYZABCDEFGH
        state:
          $ref: "#/components/schemas/SandboxState"
        runtime:
          $ref: "#/components/schemas/RuntimeType"
        image:
          type: string
          example: alpine-linux-6.1
        vcpus:
          type: integer
          example: 2
        memory_mb:
          type: integer
          example: 512
        idle_timeout_secs:
          type: integer
          example: 300
        ports:
          type: array
          items:
            $ref: "#/components/schemas/PortBinding"
        created_at:
          type: string
          format: date-time
          example: "2026-06-09T12:00:00Z"
        last_activity_at:
          type: string
          format: date-time
          example: "2026-06-09T12:05:00Z"
        labels:
          type: object
          additionalProperties:
            type: string
        failure:
          $ref: "#/components/schemas/FailureInfo"
        request_id:
          type: string
          example: req_01JXYZ123456789
    PortBinding:
      type: object
      properties:
        guest_port:
          type: integer
          example: 8080
        host_port:
          type: integer
          example: 32001
        host_address:
          type: string
          example: 127.0.0.1
    FailureInfo:
      type: object
      properties:
        code:
          $ref: "#/components/schemas/FailureCode"
        message:
          type: string
          example: "Guest agent did not respond within 300s boot timeout"
        retryable:
          type: boolean
          example: true
        occurred_at:
          type: string
          format: date-time
          example: "2026-06-09T12:10:15Z"
    Operation:
      type: object
      properties:
        operation_id:
          type: string
          pattern: "^op_[a-zA-Z0-9]+$"
          example: op_01JXYZ123456789
        sandbox_id:
          type: string
          pattern: "^sbx_[a-zA-Z0-9]+$"
          description: Present when the operation created a sandbox
          example: sbx_01JXYZABCDEFGH
        action:
          type: string
          enum:
            - create
            - destroy
            - suspend
            - resume
            - fork
          example: create
        status:
          $ref: "#/components/schemas/OperationStatus"
        state:
          $ref: "#/components/schemas/SandboxState"
          description: Current lifecycle state of the sandbox
        result:
          description: |
            The final sandbox resource when the operation is `completed`.
            Absent while `pending` or `in_progress`.
          allOf:
            - $ref: "#/components/schemas/Sandbox"
        error:
          description: Error information when the operation is `failed`
          allOf:
            - $ref: "#/components/schemas/Error"
        status_url:
          type: string
          format: uri-reference
          example: /v2/operations/op_01JXYZ123456789
        created_at:
          type: string
          format: date-time
          example: "2026-06-09T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-06-09T12:00:05Z"
        request_id:
          type: string
          example: req_01JXYZ123456789
    SandboxList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Sandbox"
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page; null on the last page
          example: cur_01JXYZ123456789
        request_id:
          type: string
          example: req_01JXYZ123456789
...
    Error:
      type: object
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: state_conflict
            message:
              type: string
              description: Human-readable error description
              example: "Sandbox must be Running to exec (current state: Suspended)"
            details:
              description: Optional structured details about the error
            request_id:
              type: string
              description: Unique identifier for this request
              example: req_01JXYZ123456789

    # ---- Streaming ----
...